syntax: the tree and its spans are region data
The fourth foundation from the image design (docs/impl/image.md §
"Landing order"), stacked on the struct foundation. Independently
valuable at runtime, and it deletes image machinery that would
otherwise have to be built and thrown away.

A `Syntax` was a Rust-heap `Box` tree: `Vec<Syntax>` children, `String`
names, `Box<Syntax>` for the quote forms, `Rc<Syntax>` for a syntax
literal, and a `Span` carrying `Option<String>` — one heap allocation
per source location, cloned on every span merge and every node copy.

That cost twice. Every macro expansion deep-cloned its template and its
arguments; the boot profile put `Syntax::clone` + `drop_in_place<Syntax>`
at ~10% of the whole run. And it made syntax undumpable: a macro *is* its
template, so an image cannot rebuild the expander without carrying
syntax, and a `Box` inside `HeapObject::Syntax` is exactly the Rust-heap
ownership the sealing rule forbids. The alternative was a side-stream
codec whose entire job would be to work around the representation.

`Syntax` is region data. A node is `Copy` POD, 64 bytes, with no `Drop`:

| Payload | Was | Is |
|---------|-----|-----|
| a name or literal | `String` | `RegionStr` |
| children | `Vec<Syntax>` | `RegionSlice<Syntax>` |
| a wrapped form | `Box<Syntax>` | `SynRef` |
| a syntax literal | `Rc<Syntax>` | `SynRef` |
| a scope set | `Vec<ScopeId>` | `RegionSlice<ScopeId>` |
| a span's file | `Option<String>` | `FileId` into a process-wide interner |

Every payload dereferences to what the owned form held — `&str`,
`&[Syntax]`, `&Syntax` — so the ~1,100 sites that read the tree match and
walk it unchanged. `Span` is `Copy` POD at 20 bytes.

Nodes are born in a `SyntaxArena`: a heap and a region on it. An instance
runs two. The **working** arena is one transient region per compilation
unit, minted by `pipeline::compile`'s `with_syntax_arena` and freed when
the unit's bytecode is built. The **template** arena is the instance's
process-root region, where `handle_defmacro` copies a macro template,
because a `MacroDef` outlives the unit that defined it. One rule keeps
every pointer valid: a node may point into its own arena or into the
template arena, and nowhere else.

`HeapObject::Syntax` holds its node inline, and `value::build::syntax`
copies the tree into the value's own region — the rule `RegionSlice`'s
module docs state for every slice-backed clone. The object is
self-contained page bytes: nothing to seal around, no cross-region edge
to record. Every builder of a syntax object obeys that rule,
`with-traits` included: a `Copy` node makes `syntax: *syntax` compile
there, and it would leave the clone's children in the source's region —
the hazard tests/elle/region-withtraits-slice-uaf.lisp pins for the other
slice-backed immutables.

docs/impl/syntax.md owns the model; image.md cites it rather than
restating it.

The hygiene protocol. Scope sets, the intro-scope flip, `scope_exempt`,
and the `datum->syntax` escape hatch behave exactly as before. The one
new obligation is that a scope walk COPIES: subtrees are shared by
pointer now, so writing through the source would stamp a tree the caller
still holds.

`SendSyntax` also stays. A region belongs to one `RegionStore`, so a tree
still crosses a thread as owned data. What changed is the refusal:
`SyntaxLiteral` was the mirror's one unsendable case because it held an
`Rc`, and it is an ordinary single-child node now.

`ClosureTemplate.syntax: Option<Rc<Syntax>>` retained a whole lambda tree
for the life of every compiled closure so that `(meta/origin f)` could
read three integers and a file name off its span. It is
`origin: Option<Span>` now — 20 bytes of POD, and one less thing an image
must seal.

Nine copies of a fifteen-arm match. Every recursive pass over the tree —
both scope walks, the deep copy, keyword learning, the `SyntaxLiteral`
scan, the epoch rewrite, the send mirror, the census — goes through one
`SyntaxKind::children` / `SyntaxKind::rebuild` pair. The change is net
smaller than what it replaces.

`config::get()` used `get_or_init`, so the first read decided the whole
process's configuration. That held while nothing allocated before
`config::init`, and this change breaks it: the pre-VM `unicode!` prescan
parses source, parsing allocates, and an allocation asks the config
whether `--stats` is on. Every flag was then discarded on the file path.
`get()` now answers a pre-init read from a separate defaults cell without
installing it, so `init` stays authoritative whatever the read order.
tests/integration/trace_boot.rs caught it and pins it.

Cold stdlib compile, minimum of nine runs, release build, same machine,
base = this change's parent:

| `<stdlib>` phase | base | this | |
|---|---:|---:|---:|
| read | 3.6 ms | 3.5 ms | |
| expand | 89.5 ms | 59.2 ms | −34% |
| analyze | 74.8 ms | 47.1 ms | −37% |
| regions | 87.0 ms | 73.5 ms | −16% |
| lower | 17.0 ms | 12.3 ms | −28% |
| emit | 36.4 ms | 26.1 ms | −28% |
| **stdlib-compile** | **324.6 ms** | **229.6 ms** | **−29%** |

Peak RSS over the same cold boot: 115,196 KiB → 90,444 KiB (−21%). The
gain runs past expansion because every HIR and LIR node carries a `Span`,
and a `Span` no longer allocates.

image.md risk item 5 projected "expansion-heavy compiles roughly a third
faster" from a prototype; the measured frontend is 29% faster end to end.

- `a_syntax_node_is_copy_pod_that_needs_no_drop`,
  `a_compounds_children_live_in_the_arenas_region`,
  `a_names_bytes_live_in_the_arenas_region` — the representation claims.
- `payloads_deref_to_str_slice_and_node` — the seam that let the read
  sites survive untouched.
- `a_copy_into_another_arena_shares_nothing_with_its_source` and
  `a_stamped_copy_does_not_disturb_its_source` — the two operations that
  cross or copy a tree.
- `a_syntax_value_owns_its_tree` — counter-factual verified: sharing
  instead of copying segfaults the moment the reader's heap drops.
- `with_traits_copies_a_syntax_tree_into_the_clones_region` — the same
  hazard at `with-traits`, where `Syntax` being `Copy` makes the mistake
  compile.
- `a_syntax_tree_round_trips_through_the_send_mirror`,
  `a_syntax_literal_crosses_the_send_mirror`.
- `a_span_is_copy_pod_with_no_rust_heap_allocation`,
  `a_span_serializes_its_file_name_not_its_id`,
  `merge_keeps_the_first_file_it_has`.
- `interning_the_same_name_twice_yields_one_id`,
  `the_absent_name_is_the_empty_name`.
- tests/integration/syntax_regions.rs —
  `a_units_syntax_arena_is_freed_with_the_unit` (counter-factual
  verified: leaving the arena live leaks one region per compile),
  `a_macro_template_outlives_the_unit_that_defined_it`,
  `a_standalone_syntax_heap_reads_and_reclaims_without_a_runtime`.

Full `make test` green with the release corpus.

+++ src/primitives/compile/spans.rs
@@ -256,7 +256,7 @@ pub(crate) fn collect_vars_in_range(
     defined: &mut BTreeSet<Binding>,
     signal: &mut Signal,
 ) {
-    if hir.span.start >= end || hir.span.end <= start {
+    if hir.span.start as usize >= end || (hir.span.end as usize) <= start {
         return;
     }
     *signal = signal.combine(hir.signal);
